refactor: rename `wrapper.ts` to `set-wrapper.ts`
