+++ packages/redis/src/set-wrapper/__tests__/wrapper.test.ts
@@ -12,7 +12,7 @@ import {
   type RedisClusterClient,
   type RedisNodeClient,
   RedisSetWrapper,
-} from "../wrapper.js";
+} from "../set-wrapper.js";
 
 // ---------------------------------------------------------------------------
 // Fixtures
